Read of uninitialized memory strongly suspected
init->len: -1094795586 (0xBEBEBEBE)
